Use ssh again for submodules.  (reverts 415fa7d)
Thanks to Laura:

On Wed, Mar 13, 2019, 11:41 Laura from Netlify <laura@netlify.com> wrote:

Hi Rob,

OK in that case you'll need to do the following:

* create a dummy user, let's call him "bot" - on github, who has access to both repositories
* remove the netlify deploy key from your repository (it's in settings -> deploy keys)
* assign the deploy key from the netlify site as the bot user's github SSH key (see below for this key)

When we check out the first repo to build, we'll fetch the submodule with the same deploy key and this will work!

The reason for this convoluted setup is that GitHub disallows an ssh key to be used in more than one location (user AND repository count as a "location"), and we can only use one ssh key per netlify site.